Add files via upload
This is my very first java script! I have been learning Java for the past 2 weeks and this game is a combination of everything I have learned so far. The game is very basic: the user declares whether they want to win via odds or evens, they pick numbers which are summed together with a random number chosen by the CPU. If the player has chosen odd and sum of the two numbers is an odd number the player receives a point. The game continues for 3 rounds. I am fully aware the code isn’t neat, efficient or structured but as I mentioned before; I have only started learning Java and this has been a perfect opportunity to play around with variables, methods, control structures and loops.